Fix broken updating
I believe that updating was broken when I changed the manifest url in the module.json file to point to the specific tagged version of the module.json file. I believe then whenever Foundry would check for updates that it would go back to that URL and see that it was the same version in the file so it wouldn't update. The manifest URL should point to an endpoint that will give the latest version such as the master branch's version of module.json. I'd like to eventually get a public endpoint on the turkeys unite server because it could be possible that the master branch isn't correct, though that really should be rare