Refactor drop methods
Organized the methods that handle the drop data from the different versions of Foundry. Added a function to normalize the data. This function takes the data from the different types of drops, the js drop event and the dropCanvasData hook from Foundry and creates an object that can be used by the method that actually handles dropping an item onto the canvas to create the token or other actions.