Create loot token
Began work on the new way of creating a token that represents loot.
Using a class named LootToken to represent this. It'll have a static
create method that will return a new instance of LootToken which will
also be added to the game setting that contains all of the loot data