Create LootHud instances for previously created loot data
In the case of reloading a world or changing scenes and coming back, we
needed a way to re-intialize the listeners on LootToken instances since
their previous Token instances would be gone. So on canvas ready, we
look through the loot token data and try and either find an already
created LootToken instance if one exists and if not then a newly created
Token instance that matches the correct IDs in the data and create a ne
LootToken instance. We then setup the listeners on the new token
instance